Move sql queries to a new object
